Final prep demo

+++ backend/routes/likes/likes.js
@@ -1,24 +1,28 @@
 require("dotenv").config();
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
 const router = require("express").Router();
 const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
 const { user, post } = new PrismaClient();
 // const { JSON } = require("express");
+const { getJWT, verifyJWT } = require("../../utils/jwt");
+const { getUserExists, getPostExists, getLikeExists } = require("../../utils/database");
 
 // Create a user like
 router.post('/createUserLike', async (req, res) => {
-
     try {
-        const { userID, postID} = req.body
-        const userExists = await prisma.User.findUnique({
-            where: { id: userID },
-        });
+        const { userID, postID, token } = req.body;
 
-        const postExists = await prisma.Post.findUnique({
-            where: { id: postID },
-        });
+        const decoded = verifyJWT(token);
+
+        if (!decoded) {
+            return res.status(400).json({
+                msg: "Invalid token"
+            });
+        }
+
+        const userExists = await getUserExists(userID, "id");
+
+        const postExists = await getPostExists(postID, "id");
 
         if (!userExists) {
             return res.status(400).json({
@@ -27,8 +31,16 @@ router.post('/createUserLike', async (req, res) => {
         } else if (!postExists) { 
             return res.status(400).json({
                 msg: "Post not found"
-            })
+            });
         } else {
+            const likeExists = await getLikeExists(postID, userID);
+
+            if (likeExists) {
+                return res.status(400).json({
+                    msg: "Like already exists"
+                });
+            }
+
             // Create a like
             const newLike = await prisma.Like.create({
                 data: {
@@ -37,31 +49,76 @@ router.post('/createUserLike', async (req, res) => {
                 }
             });
 
+            const updatePost = await prisma.Post.update({
+                where: { id: postID },
+                data: {
+                    likeCount: postExists.likeCount + 1
+                }
+            });
+
             res.json(newLike);
         }
     } catch (err) {
-        res.status(500).send({ msg: err })
+        console.log(err);
+        res.status(500).send({ msg: err });
     }
-
 });
 
 // Remove a user like
 router.delete('/removeUserLike', async (req, res) => { 
-    //try {
-        console.log(req.body.userID, req.body.postID);
-        const deleteLike = await prisma.Like.delete({
-            where: { 
-                postID_userID: {
-                    postID: req.body.postID,
-                    userID: req.body.userID,
-                },
+    try {
+
+        const { userID, postID, token } = req.body;
+
+        const decoded = verifyJWT(token);
+
+        if (!decoded) {
+            return res.status(400).json({
+                msg: "Invalid token"
+            });
+        }
+
+        const userExists = await getUserExists(userID, "id");
+
+        const postExists = await getPostExists(postID, "id");
+
+        if (!userExists) {
+            return res.status(400).json({
+                msg: "User not found"
+            });
+        } else if (!postExists) { 
+            return res.status(400).json({
+                msg: "Post not found"
+            });
+        } else {
+            const deleteLike = await prisma.Like.delete({
+                where: { 
+                    postID_userID: {
+                        postID: postID,
+                        userID: userID,
+                    },
+                }
+            });
+
+            if (!deleteLike) {
+                return res.status(400).json({
+                    msg: "Like not found"
+                });
             }
-        });
-        console.log(deleteLike);
-        res.status(200).send({ msg: "Deleted a user like" });
-    //} catch (err) {
-        //res.status(500).send(err);
-    //}
+    
+            const updatePost = await prisma.Post.update({
+                where: { id: postID },
+                data: {
+                    likeCount: postExists.likeCount - 1
+                }
+            });
+    
+            res.status(200).send({ msg: "Deleted a user like" });
+        }
+    } catch (err) {
+        console.log(err);
+        res.status(500).send(err);
+    }
 });
 
 // Get user like status
@@ -78,6 +135,7 @@ router.get('/getUserLike', async (req, res) => {
 
         res.json(likeStatus);
     } catch (err) {
+        console.log(err);
         res.status(500).send(err);
     }
 });
@@ -91,6 +149,7 @@ router.get('/getAllUserLikes', async (req, res) => {
 
         res.json(allLikes);
     } catch (err) {
+        console.log(err);
         res.status(500).send(err);
     }
 });
