+++ backend/routes/playlists/playlists.js
@@ -1,73 +1,96 @@
 require("dotenv").config();
-const bcrypt = require('bcryptjs');
-const jwt = require('jsonwebtoken');
 const router = require("express").Router();
 const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
 const { user, post } = new PrismaClient();
 // const { JSON } = require("express");
+const { getJWT, verifyJWT } = require("../../utils/jwt");
+const { getUserExists, getPostExists, getPlaylistExists} = require("../../utils/database");
+const multer  = require('multer')
+const upload = multer()
 
 // Create a new playlist
-router.post('/createPlaylist', async (req, res) => {
+
+//Thumbnail is a file upload is here
+router.post('/createPlaylist', upload.single('thumbnail'), async (req, res) => {
 
     try {
-        const { name, userID } = req.body;
+        const { name, userID, token, thumbnail} = req.body;
+        const decoded = verifyJWT(token);
 
-        const userExists = await prisma.User.findUnique({
-            where: { id: userID }
-        });
+        if (!decoded) {
+            return res.status(400).json({
+                msg: "Invalid token"
+            });
+        }
+
+        const userExists = await getUserExists(userID, "id");
 
         if (!userExists) {
             return res.status(400).json({
                 msg: "User not found"
-            })
+            });
         } else {
-
             const newPlaylist = await prisma.Playlist.create({
                 data: {
                     name: name,
                     userID: userID,
+                    thumbnail: thumbnail
                 }
             });
-
+      
+            // console.log(profilePicture);
             res.json(newPlaylist);
         }
     } catch (err) {
-        res.status(500).send({ msg: err })
+        console.log(err);
+        res.status(500).send({ msg: err });
     }
-
 });
 
 
 // Get all playlists
 router.get('/getAllPlaylists', async (req, res) => {
-
     try {
         const playlists = await prisma.Playlist.findMany();
         res.json(playlists);
+    } catch (err) {
+        console.log(err);
+        res.status(500).send({ msg: err });
     }
-    catch (err) {
-        res.status(500).send({ msg: err })
-    }
+});
 
+// Get all playlists for a user
+router.get('/getUserPlaylists', async (req, res) => {
+    const userID = req.query.userID;
+
+    try {
+        const playlists = await prisma.Playlist.findMany({
+            where: {
+                userID: userID
+            }
+        });
+        res.json(playlists);
+    } catch (err) {
+        console.log(err);
+        res.status(500).send({ msg: err });
+    }
 });
 
 // Get playlist by ID
 router.get('/getPlaylistByID', async (req, res) => {
     try {
-        const playlist = await prisma.Playlist.findUnique({
-            where: { id: req.query.id }
-        });
+        const playlistExists = await getPlaylistExists(req.query.id, "id");
 
-        if (!playlist) {
+        if (!playlistExists) {
             return res.status(400).json({
                 msg: "Playlist does not exist."
             });
         }
 
-        res.json(playlist);
-    }
-    catch (err) {
+        res.json(playlistExists);
+    } catch (err) {
+        console.log(err);
         res.status(500).send({ msg: err })
     }
 });
@@ -91,8 +114,8 @@ router.get('/getPlaylistsByPostID', async (req, res) => {
         }
 
         res.json(playlists);
-    }
-    catch (err) {
+    } catch (err) {
+        console.log(err);
         res.status(500).send({ msg: err })
     }
 
@@ -101,11 +124,9 @@ router.get('/getPlaylistsByPostID', async (req, res) => {
 // Get all posts in a playlist
 router.get('/getPostsByPlaylistID', async (req, res) => {
     try {
-        const playlist = await prisma.Playlist.findUnique({
-            where: { id: req.query.id }
-        });
+        const playlistExists = await getPlaylistExists(req.query.id, "id");
 
-        if (!playlist) {
+        if (!playlistExists) {
             return res.status(400).json({
                 msg: "Playlist does not exist."
             });
@@ -118,9 +139,9 @@ router.get('/getPostsByPlaylistID', async (req, res) => {
         }
 
         res.json(playlist);
-    }
-    catch (err) {
-        res.status(500).send({ msg: err })
+    } catch (err) {
+        console.log(err);
+        res.status(500).send({ msg: err });
     }
 });
 
@@ -128,12 +149,22 @@ router.get('/getPostsByPlaylistID', async (req, res) => {
 router.delete('/deletePlaylist', async (req, res) => {
 
     try {
+        const decoded = verifyJWT(req.body.token);
+
+        if (!decoded) {
+            return res.status(400).json({
+                msg: "Invalid token"
+                });
+        }
+
         const deletePlaylist = await prisma.Playlist.delete({
             where: { id: req.body.id }
-        })
+        });
+
         res.status(200).send({ msg: "Deleted a user playlist" });
     }
     catch (err) {
+        console.log(err);
         res.status(500).send(err);
     }
 
@@ -143,15 +174,19 @@ router.delete('/deletePlaylist', async (req, res) => {
 // Put post in playlist
 router.post('/addPostToPlaylist', async (req, res) => {
     try {
+        const { playlistID, postID, token } = req.body;
         
-        const { playlistID, postID } = req.body;
-        const playlistExists = await prisma.Playlist.findUnique({
-            where: { id: playlistID },
-        });
+        const decoded = verifyJWT(token);
 
-        const postExists = await prisma.Post.findUnique({
-            where: { id: postID },
-        });
+        if (!decoded) {
+            return res.status(400).json({
+                msg: "Invalid token"
+            });
+        }
+
+        const playlistExists = await getPlaylistExists(playlistID, "id");
+
+        const postExists = await getPostExists(postID, "id");
 
         if (!playlistExists) {
             return res.status(400).json({
@@ -172,14 +207,24 @@ router.post('/addPostToPlaylist', async (req, res) => {
         }
     }
     catch (err) {
-        res.status(500).send({ msg: err })
+        console.log(err);
+        res.status(500).send({ msg: err });
     }
 });
 
 // Remove a post from a playlist
 router.delete('/removePostFromPlaylist', async (req, res) => {
     try {
-        const { postID, playlistID } = req.body;
+        const { postID, playlistID, token } = req.body;
+
+        const decoded = verifyJWT(token);
+
+        if (!decoded) {
+            return res.status(400).json({
+                msg: "Invalid token"
+            });
+        }
+
         const removePost = await prisma.PlaylistPost.delete({
             where: {
                 postID_playlistID: {
@@ -191,31 +236,37 @@ router.delete('/removePostFromPlaylist', async (req, res) => {
         res.status(200).send({ msg: "Removed a post from a playlist" });
     }
     catch (err) {
+        console.log(err);
         res.status(500).send(err);
     }
-
 });
 
 // TODO : Update a playlist
 router.put('/updatePlaylist', async (req, res) => {
-
     try {
-        const { id, name, thumbnail} = req.body
-        const updateUser = await prisma.Playlist.update({
+        const { id, name, thumbnail, token} = req.body;
+
+        const decoded = verifyJWT(token);
+
+        if (!decoded) {
+            return res.status(400).json({
+                msg: "Invalid token"
+                });
+        }
+
+        const updatePlaylist = await prisma.Playlist.update({
             where: { id },
             data: {
                 name: name,
                 thumbnail: thumbnail
             }
-        })
+        });
         //   res.status(200).send({msg: "Updated OK"});
-        res.json(updateUser);
-    }
-
-    catch (err) {
+        res.json(updatePlaylist);
+    } catch (err) {
+        console.log(err);
         res.status(500).send(err);
     }
-
 });
 
 module.exports = router;
